4.15-4.23

+++ Osa4/blogs/controllers/blogs.js
@@ -1,34 +1,46 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog.js')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async(request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username:1, name:1, id:1})
   response.json(blogs)
   })
   
-blogsRouter.post('/', async(request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async(request, response) => {
     const body = request.body
 
     if( !body.title || !body.url) {
       response.status(400).end()
     }
-    
+
     else {
+    const user = request.user
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes | 0
+        likes: body.likes | 0,
+        user: user._id
     })
 
     const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
+    user.blogs = user.blogs.concat(savedBlog.id)
+    await user.save()
+    response.json(savedBlog)
   }
   })
 
-blogsRouter.delete('/:id', async(request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async(request, response) => {
+
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+  
+  if(blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else response.status(400) 
 })
 
 blogsRouter.put('/:id', async(request, response) => {
